refactor: merge types.go into volume_builder.go
Consolidate ListenerScope type and ListenerStorageClassPtr helper
into volume_builder.go alongside the other types and constants,
removing the unnecessary types.go file.
