--- pkg/listener/types.go
@@ -1,35 +1,0 @@
-/*
-Copyright 2024 ZNCDataDev.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package listener
-
-import "k8s.io/utils/ptr"
-
-// ListenerScope defines the scope of listener volumes provisioned by listener-operator.
-type ListenerScope string
-
-const (
-	// ListenerScopeNode limits listener discovery to the node level.
-	ListenerScopeNode ListenerScope = "Node"
-	// ListenerScopeCluster enables listener discovery across the cluster.
-	ListenerScopeCluster ListenerScope = "Cluster"
-)
-
-// ListenerStorageClassPtr returns a pointer to the ListenerStorageClass constant.
-// Useful for Kubernetes PVC spec fields that require *string.
-func ListenerStorageClassPtr() *string {
-	return ptr.To(ListenerStorageClass)
-}
